Handle months with accents in TNTTest
The build broke when we reached "décembre" with an accent, because Date.parse() cannot parse a string with an accent.

By asserting presence of french months' in text string, the tests are more reliable.

Also use the instance variable @tnt like in ever other method in the test class.